# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Silo Feed Tracker API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Silo {
  id: number;
  name: string;
  /** @nullable */
  defaultFeedType: string | null;
  createdAt: string;
}

export interface CreateSiloBody {
  name: string;
  /** @nullable */
  defaultFeedType?: string | null;
}

export interface Reading {
  id: number;
  siloId: number;
  siloName: string;
  feedType: string;
  amountRemaining: number;
  unit: string;
  /** @nullable */
  notes: string | null;
  readingDate: string;
  createdAt: string;
}

export interface CreateReadingBody {
  siloId: number;
  feedType: string;
  amountRemaining: number;
  unit: string;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  readingDate?: string | null;
}

export interface SiloSummary {
  siloId: number;
  siloName: string;
  feedType: string;
  amountRemaining: number;
  unit: string;
  readingDate: string;
  lastReadingId: number;
}

export interface OnedriveStatus {
  connected: boolean;
  /** @nullable */
  fileId: string | null;
  /** @nullable */
  fileName: string | null;
}

export type ListReadingsParams = {
  siloId?: number;
  limit?: number;
};