# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Silo Feed Tracker API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary List all silos
 */
export const ListSilosResponseItem = zod.object({
  id: zod.number(),
  name: zod.string(),
  defaultFeedType: zod.string().nullable(),
  createdAt: zod.string(),
});
export const ListSilosResponse = zod.array(ListSilosResponseItem);

/**
 * @summary Create a silo
 */
export const CreateSiloBody = zod.object({
  name: zod.string(),
  defaultFeedType: zod.string().nullish(),
});

/**
 * @summary Update a silo
 */
export const UpdateSiloParams = zod.object({
  id: zod.coerce.number(),
});

export const UpdateSiloBody = zod.object({
  name: zod.string(),
  defaultFeedType: zod.string().nullish(),
});

export const UpdateSiloResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  defaultFeedType: zod.string().nullable(),
  createdAt: zod.string(),
});

/**
 * @summary Delete a silo
 */
export const DeleteSiloParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary List all readings (newest first)
 */
export const ListReadingsQueryParams = zod.object({
  siloId: zod.coerce.number().optional(),
  limit: zod.coerce.number().optional(),
});

export const ListReadingsResponseItem = zod.object({
  id: zod.number(),
  siloId: zod.number(),
  siloName: zod.string(),
  feedType: zod.string(),
  amountRemaining: zod.number(),
  unit: zod.string(),
  notes: zod.string().nullable(),
  readingDate: zod.string(),
  createdAt: zod.string(),
});
export const ListReadingsResponse = zod.array(ListReadingsResponseItem);

/**
 * @summary Record a silo reading
 */
export const CreateReadingBody = zod.object({
  siloId: zod.number(),
  feedType: zod.string(),
  amountRemaining: zod.number(),
  unit: zod.string(),
  notes: zod.string().nullish(),
  readingDate: zod.string().nullish(),
});

/**
 * @summary Get the latest reading for each silo
 */
export const GetReadingsSummaryResponseItem = zod.object({
  siloId: zod.number(),
  siloName: zod.string(),
  feedType: zod.string(),
  amountRemaining: zod.number(),
  unit: zod.string(),
  readingDate: zod.string(),
  lastReadingId: zod.number(),
});
export const GetReadingsSummaryResponse = zod.array(
  GetReadingsSummaryResponseItem,
);

/**
 * @summary Delete a reading
 */
export const DeleteReadingParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary Check if OneDrive is connected
 */
export const GetOnedriveStatusResponse = zod.object({
  connected: zod.boolean(),
  fileId: zod.string().nullable(),
  fileName: zod.string().nullable(),
});