# lib/api-zod/src/generated/types/createReadingBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Silo Feed Tracker API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateReadingBody {
  siloId: number;
  feedType: string;
  amountRemaining: number;
  unit: string;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  readingDate?: string | null;
}