# lib/api-zod/src/generated/types/onedriveStatus.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Silo Feed Tracker API
 * OpenAPI spec version: 0.1.0
 */

export interface OnedriveStatus {
  connected: boolean;
  /** @nullable */
  fileId: string | null;
  /** @nullable */
  fileName: string | null;
}