# lib/api-zod/src/generated/types/reading.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Silo Feed Tracker API
 * OpenAPI spec version: 0.1.0
 */

export interface Reading {
  id: number;
  siloId: number;
  siloName: string;
  feedType: string;
  amountRemaining: number;
  unit: string;
  /** @nullable */
  notes: string | null;
  readingDate: string;
  createdAt: string;
}