# lib/api-zod/src/generated/types/silo.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Silo Feed Tracker API
 * OpenAPI spec version: 0.1.0
 */

export interface Silo {
  id: number;
  name: string;
  /** @nullable */
  defaultFeedType: string | null;
  createdAt: string;
}